rename structured to octagon file
