+++ src/pygeon/grids/octagon.py
